manage: Refactor and fix /packages/open_vulns query
- Completely rewritten the query code to improve readability.
- Fixed query data display after filtering and sorting.
- Slightly improved performance.

Warning! If you find it strange that the query doesn't always return the
desired images, please know that this is not a bug. The issue is that
image searches are performed strictly by package hashes, using the
PackagesVulnerabilityStatus table, which stores data for the current
state of the repository (only for packages as of today).
The vast majority of images contain packages that are no longer the
latest.